requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
onnxruntime==1.16.3
pillow==10.2.0
numpy==1.23.5
python-multipart==0.0.9

